j'ai fais quelque modif, notament maintenant ca lis un fichier gts, avec prise en compte des # pour des ceventuezl commentaire, j'ai changer quelque truc de maxime. Reste a faire un affichage via la classe correspondante, mais le probleme est que draw polygon demande un int et que les coord sont en float, une conversion est donc necessaire...
J'inclue un fichier gts qui peut vous plaires ;)
Noraj de mon français !

+++ taupe_models1/src/Affichage.java
@@ -29,15 +29,18 @@ public class Affichage extends JPanel{
     public void paintComponent(Graphics g){
         Collections.sort(listTriangle);
         for(Triangle i : listTriangle){
-            int[] x, y;
-            x = new int[3];
-            y = new int[3];
+            float[] x, y;
+            x = new float[3];
+            y = new float[3];
             for(int idx = 0; idx < 3; idx++){
                 x[idx] = i.point.get(idx).x;
                 y[idx] = i.point.get(idx).y;
             }
             g.setColor(Color.GRAY);
+            //a trouver une astuce
+            /*
             g.drawPolygon(x, y, 3);
+                    */
         }
     }
 }
