+++ taupe_models1/src/Point.java
@@ -11,11 +11,11 @@
  * @author flo
  */
 public class Point {
-    int x;
-    int y;
-    int z;
+    float x;
+    float y;
+    float z;
     
-    public Point(int x, int y, int z){
+    public Point(float x, float y, float z){
         this.x = x;
         this.y = y;
         this.z = z;
